% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHadamard.R
\name{CHadamard}
\alias{CHadamard}
\title{controlled Hadamard Gate}
\usage{
CHadamard(q)
}
\arguments{
\item{q}{The input}
}
\value{
k
}
\description{
This function applies a controlled Hadamard gate om the input
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Initialze global variables
init()
CHadamard(q00_)
CHadamard(I4)

}
\author{
Tinniam V Ganesh
}
\references{
\url{https://quantumexperience.ng.bluemix.net/}\cr
\url{https://gigadom.wordpress.com/2016/06/23/introducing-qcsimulator-a-5-qubit-quantum-computing-simulator-in-r/}\cr
}
\seealso{
\code{\link{CNOT2_01}}\cr
\code{\link{PauliX}}\cr
}

