% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNOT3_02.R
\name{CNOT3_02}
\alias{CNOT3_02}
\title{3 qubit CNOT gate (control-0,target-2)}
\usage{
CNOT3_02(a)
}
\arguments{
\item{a}{The input}
}
\value{
result
The result of applying the CNOT3_02 gate
}
\description{
This function applies a CNOT gate to 3 qubits. The qubits start from 0,1,2,3,4. Here
control is qubit 0 and target is qubit 2
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Initialze global variables
init()
CNOT3_02(q111_)
CNOT3_02(I8)

}
\author{
Tinniam V Ganesh
}
\references{
\url{https://quantumexperience.ng.bluemix.net/}\cr
\url{https://gigadom.wordpress.com/2016/06/23/introducing-qcsimulator-a-5-qubit-quantum-computing-simulator-in-r/}\cr
}
\seealso{
\code{\link{CNOT2_10}}\cr
\code{\link{PauliX}}\cr
\code{\link{measurement}}\cr
\code{\link{plotMeasurement}}\cr
\code{\link{CNOT5_03}}\cr
\code{\link{CNOT3_12}}\cr
}

