% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNOT4_32.R
\name{CNOT4_32}
\alias{CNOT4_32}
\title{4 qubit CNOT gate (control-3,target-2)}
\usage{
CNOT4_32(a)
}
\arguments{
\item{a}{The input}
}
\value{
result
The result of applying the CNOT4_32 gate
}
\description{
This function applies a CNOT gate to 4 qubits. The qubits start from 0,1,2,3,4. Here
control is qubit 3 and target is qubit 2
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Initialze global variables
init()
CNOT4_32(q1001_)
CNOT4_32(I16)

}
\author{
Tinniam V Ganesh
}
\references{
\url{https://quantumexperience.ng.bluemix.net/}\cr
\url{https://gigadom.wordpress.com/2016/06/23/introducing-qcsimulator-a-5-qubit-quantum-computing-simulator-in-r/}\cr
}
\seealso{
\code{\link{CNOT2_10}}\cr
\code{\link{PauliX}}\cr
\code{\link{measurement}}\cr
\code{\link{plotMeasurement}}\cr
\code{\link{CNOT5_03}}\cr
\code{\link{CNOT3_12}}\cr
}

