% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNOT5_04.R
\name{CNOT5_04}
\alias{CNOT5_04}
\title{5 qubit CNOT gate (control-0,target-4)}
\usage{
CNOT5_04(a)
}
\arguments{
\item{a}{The input}
}
\value{
result
The result of applying the CNOT5_04 gate
}
\description{
This function applies a CNOT gate to 5 qubits. The qubits start from 0,1,2,3,4. Here
control is qubit 0 and target is qubit 3
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Initialze global variables
init()
CNOT5_04(q10010_)
CNOT5_04(I32)

}
\author{
Tinniam V Ganesh
}
\references{
\url{https://quantumexperience.ng.bluemix.net/}\cr
\url{https://gigadom.wordpress.com/2016/06/23/introducing-qcsimulator-a-5-qubit-quantum-computing-simulator-in-r/}\cr
}
\seealso{
\code{\link{CNOT2_10}}\cr
\code{\link{PauliX}}\cr
\code{\link{measurement}}\cr
\code{\link{plotMeasurement}}\cr
\code{\link{CNOT5_03}}\cr
\code{\link{CNOT3_12}}\cr
}

