% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPauliX.R
\name{CPauliX}
\alias{CPauliX}
\title{Controlled Pauli X gate}
\usage{
CPauliX(q)
}
\arguments{
\item{q}{The input}
}
\value{
2
The result of applying the CPauliX gate
}
\description{
This function applies a controlled Pauli X gate on its input
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Initialze global variables
init()
CPauliX(q11_)
CPauliX(I4)


}
\author{
Tinniam V Ganesh
}
\references{
\url{https://quantumexperience.ng.bluemix.net/}\cr
\url{https://gigadom.wordpress.com/2016/06/23/introducing-qcsimulator-a-5-qubit-quantum-computing-simulator-in-r/}\cr
}
\seealso{
\code{\link{CNOT2_10}}\cr
\code{\link{PauliX}}\cr
\code{\link{measurement}}\cr
\code{\link{plotMeasurement}}\cr
\code{\link{CNOT5_03}}\cr
\code{\link{CNOT5_13}}\cr
}

