% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPauliZ.R
\name{CPauliZ}
\alias{CPauliZ}
\title{Controlled Pauli Z gate}
\usage{
CPauliZ(q)
}
\arguments{
\item{q}{The input}
}
\value{
2
The result of applying the CPauliZ gate
}
\description{
This function applies a controlled Pauli Z gate on its input
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Initialze global variables
init()
CPauliZ(q11_)
CPauliZ(I4)


}
\author{
Tinniam V Ganesh
}
\references{
\url{https://quantumexperience.ng.bluemix.net/}\cr
\url{https://gigadom.wordpress.com/2016/06/23/introducing-qcsimulator-a-5-qubit-quantum-computing-simulator-in-r/}\cr
}
\seealso{
\code{\link{CNOT2_10}}\cr
\code{\link{PauliX}}\cr
\code{\link{measurement}}\cr
\code{\link{plotMeasurement}}\cr
\code{\link{CNOT5_03}}\cr
\code{\link{CNOT4_13}}\cr
}

