% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DotProduct.R
\name{DotProduct}
\alias{DotProduct}
\title{Dot product of 2 vectors}
\usage{
DotProduct(a,b)
}
\arguments{
\item{a}{Vector 1}

\item{b}{Vector 2}
}
\value{
result
The result of dot product
}
\description{
This function performs a dot product of 2 vectors
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Initialze global variables
init()
DotProduct(Hadamard(I2),q1_)
DotProduct(CNOT2_01(I4),q01_)

}
\author{
Tinniam V Ganesh
}
\references{
\url{https://quantumexperience.ng.bluemix.net/}\cr
\url{https://gigadom.wordpress.com/2016/06/23/introducing-qcsimulator-a-5-qubit-quantum-computing-simulator-in-r/}\cr
}
\seealso{
\code{\link{CNOT2_10}}\cr
\code{\link{PauliX}}\cr
\code{\link{measurement}}\cr
\code{\link{plotMeasurement}}\cr
\code{\link{CNOT5_03}}\cr
\code{\link{CNOT4_13}}\cr
}

