% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innerProduct.R
\name{innerProduct}
\alias{innerProduct}
\title{Inner product of 2 vectors and computes the angle between vectors}
\usage{
innerProduct(a,b)
}
\arguments{
\item{a}{Vector 1}

\item{b}{Vector 2}
}
\value{
theta
The angle between the vectors
}
\description{
This function performs a inner  product of 2 vectors and outputs the angle between
vectors
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Initialze global variables
init()
phi = matrix(c(1/2,sqrt(3)/2),nrow=2,ncol=1)
si = matrix(c(1/sqrt(2),1/sqrt(2)),nrow=2,ncol=1)
innerProduct(phi,si)


}
\author{
Tinniam V Ganesh
}
\references{
\url{https://quantumexperience.ng.bluemix.net/}\cr
\url{https://gigadom.wordpress.com/2016/06/23/introducing-qcsimulator-a-5-qubit-quantum-computing-simulator-in-r/}\cr
}
\seealso{
\code{\link{CNOT2_10}}\cr
\code{\link{PauliX}}\cr
\code{\link{measurement}}\cr
\code{\link{plotMeasurement}}\cr
\code{\link{CNOT5_03}}\cr
\code{\link{CNOT5_13}}\cr
}

