% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement.R
\name{measurement}
\alias{measurement}
\title{Computes the square of the modulus}
\usage{
measurement(a)
}
\arguments{
\item{a}{The vector}
}
\value{
x
The square of the modulus of the vector
}
\description{
This function computes the square of the amplitude of the vectors
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Initialze global variables
init()
measurement(TGate(PauliX(I2)))


}
\author{
Tinniam V Ganesh
}
\references{
\url{https://quantumexperience.ng.bluemix.net/}\cr
\url{https://gigadom.wordpress.com/2016/06/23/introducing-qcsimulator-a-5-qubit-quantum-computing-simulator-in-r/}\cr
}
\seealso{
\code{\link{CNOT2_10}}\cr
\code{\link{PauliX}}\cr
\code{\link{measurement}}\cr
\code{\link{plotMeasurement}}\cr
\code{\link{CNOT5_03}}\cr
\code{\link{CNOT5_13}}\cr
}

