% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMeasurement.R
\name{plotMeasurement}
\alias{plotMeasurement}
\title{This function plots the result of a measurement}
\usage{
plotMeasurement(a)
}
\arguments{
\item{a}{The vector}
}
\description{
This function plots the output of a quantum circuit
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Initialze global variables
init()
plotMeasurement(measurement(TensorProd(Hadamard(I2),Hadamard(I2))))


}
\author{
Tinniam V Ganesh
}
\references{
\url{https://quantumexperience.ng.bluemix.net/}\cr
\url{https://gigadom.wordpress.com/2016/06/23/introducing-qcsimulator-a-5-qubit-quantum-computing-simulator-in-r/}\cr
}
\seealso{
\code{\link{CNOT2_10}}\cr
\code{\link{PauliX}}\cr
\code{\link{measurement}}\cr
\code{\link{plotMeasurement}}\cr
\code{\link{CNOT5_03}}\cr
\code{\link{CNOT4_13}}\cr
}

