% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RotationGate.R
\name{RotationGate}
\alias{RotationGate}
\title{This fucntion applies the rotation gate}
\usage{
RotationGate(t,a)
}
\arguments{
\item{t}{The angle to rotate}

\item{a}{The vector}
}
\value{
result
The result of applying the Rotation gate
}
\description{
This function applies the rotation gate on its input through an angle 't'
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Initialze global variables
init()
RotationGate(30,q0_)


}
\author{
Tinniam V Ganesh
}
\references{
\url{https://quantumexperience.ng.bluemix.net/}\cr
\url{https://gigadom.wordpress.com/2016/06/23/introducing-qcsimulator-a-5-qubit-quantum-computing-simulator-in-r/}\cr
}
\seealso{
\code{\link{CNOT2_10}}\cr
\code{\link{PauliX}}\cr
\code{\link{measurement}}\cr
\code{\link{plotMeasurement}}\cr
\code{\link{CNOT5_03}}\cr
\code{\link{CNOT4_13}}\cr
}

