% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGate.R
\name{SGate}
\alias{SGate}
\title{Apply a SGate}
\usage{
SGate(a)
}
\arguments{
\item{a}{The input}
}
\value{
result
The result of applying the SGate gate
}
\description{
This function applies a  SGate on its input
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Initialze global variables
init()
SGate(q1_)
SGate(Hadamard(I2))


}
\author{
Tinniam V Ganesh
}
\references{
\url{https://quantumexperience.ng.bluemix.net/}\cr
\url{https://gigadom.wordpress.com/2016/06/23/introducing-qcsimulator-a-5-qubit-quantum-computing-simulator-in-r/}\cr
}
\seealso{
\code{\link{CNOT2_10}}\cr
\code{\link{PauliX}}\cr
\code{\link{measurement}}\cr
\code{\link{plotMeasurement}}\cr
\code{\link{CNOT5_03}}\cr
\code{\link{CNOT4_13}}\cr
}

