% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWAPQ0Q1.R
\name{SWAPQ0Q1}
\alias{SWAPQ0Q1}
\title{SWAP Q0 Q1}
\usage{
SWAPQ0Q1(q)
}
\arguments{
\item{q}{The input}
}
\description{
This function swaps q0 and q1
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Initialze global variables
init()
SWAPQ0Q1(q110_)
SWAPQ0Q1(q010_)


}
\author{
Tinniam V Ganesh
}
\references{
\url{https://quantumexperience.ng.bluemix.net/}\cr
\url{https://gigadom.wordpress.com/2016/06/23/introducing-qcsimulator-a-5-qubit-quantum-computing-simulator-in-r/}\cr
}
\seealso{
\code{\link{CNOT2_10}}\cr
\code{\link{PauliX}}\cr
\code{\link{measurement}}\cr
\code{\link{plotMeasurement}}\cr
\code{\link{CNOT5_03}}\cr
\code{\link{CNOT4_13}}\cr
}

