% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T1Gate.R
\name{T1Gate}
\alias{T1Gate}
\title{Apply a T1Gate}
\usage{
T1Gate(a)
}
\arguments{
\item{a}{The input}
}
\value{
result
The result of applying the T1Gate gate
}
\description{
This function applies a  T1Gate on its input
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Initialze global variables
init()
T1Gate(q1_)
T1Gate(SGate(Hadamard(I2)))


}
\author{
Tinniam V Ganesh
}
\references{
\url{https://quantumexperience.ng.bluemix.net/}\cr
\url{https://gigadom.wordpress.com/2016/06/23/introducing-qcsimulator-a-5-qubit-quantum-computing-simulator-in-r/}\cr
}
\seealso{
\code{\link{CNOT2_10}}\cr
\code{\link{PauliX}}\cr
\code{\link{measurement}}\cr
\code{\link{plotMeasurement}}\cr
\code{\link{CNOT5_03}}\cr
\code{\link{CNOT3_12}}\cr
}

